Move http client implementations into http package. #394
