+++ sdk-http-apache-async/src/main/java/io/sphere/sdk/http/CompletableFutureCallbackAdapter.java
@@ -1,4 +1,4 @@
-package io.sphere.sdk.client;
+package io.sphere.sdk.http;
 
 import org.apache.http.concurrent.FutureCallback;
