+++ sdk-http-ning-1_9/src/main/java/io/sphere/sdk/http/NingHttpClientAdapterImpl.java
@@ -1,14 +1,16 @@
-package io.sphere.sdk.client;
+package io.sphere.sdk.http;
 
 import com.ning.http.client.*;
-import io.sphere.sdk.http.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.Optional;
-import java.util.concurrent.*;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.Executor;
+import java.util.concurrent.ForkJoinPool;
 
 final class NingHttpClientAdapterImpl implements NingHttpClientAdapter {
     private static final Logger LOGGER = LoggerFactory.getLogger(NingHttpClientAdapterImpl.class);
@@ -20,27 +22,21 @@ final class NingHttpClientAdapterImpl implements NingHttpClientAdapter {
     }
 
     @Override
-    public CompletionStage<HttpResponse> execute(final HttpRequest httpRequest) {
+    public CompletableFuture<HttpResponse> execute(final HttpRequest httpRequest) {
         if (LOGGER.isDebugEnabled()) {
             LOGGER.debug("executing " + httpRequest);
         }
         final Request request = asNingRequest(httpRequest);
-        try {
-            final CompletionStage<Response> future = wrap(asyncHttpClient.executeRequest(request));
-            return future.thenApply((Response response) -> {
-                final byte[] responseBodyAsBytes = getResponseBodyAsBytes(response);
-                Optional<byte[]> body = responseBodyAsBytes.length > 0 ? Optional.of(responseBodyAsBytes) : Optional.empty();
-                final HttpResponse httpResponse = HttpResponse.of(response.getStatusCode(), body, Optional.of(httpRequest), HttpHeaders.of(response.getHeaders()));
-                if (LOGGER.isDebugEnabled()) {
-                    LOGGER.debug("response " + httpResponse);
-                }
-                return httpResponse;
-            });
-        } catch (final IOException e) {
-            final CompletableFuture<HttpResponse> future = new CompletableFuture<>();
-            future.completeExceptionally(new HttpException(e));
-            return future;
-        }
+        final CompletableFuture<Response> future = wrap(asyncHttpClient.executeRequest(request));
+        return future.thenApply((Response response) -> {
+            final byte[] responseBodyAsBytes = getResponseBodyAsBytes(response);
+            Optional<byte[]> body = responseBodyAsBytes.length > 0 ? Optional.of(responseBodyAsBytes) : Optional.empty();
+            final HttpResponse httpResponse = HttpResponse.of(response.getStatusCode(), body, Optional.of(httpRequest), HttpHeaders.of(response.getHeaders()));
+            if (LOGGER.isDebugEnabled()) {
+                LOGGER.debug("response " + httpResponse);
+            }
+            return httpResponse;
+        });
     }
 
     private byte[] getResponseBodyAsBytes(final Response response) {
@@ -66,7 +62,7 @@ final class NingHttpClientAdapterImpl implements NingHttpClientAdapter {
             } else if (body instanceof FileHttpRequestBody) {
                 builder.setBody(((FileHttpRequestBody) body).getFile());
             } else if (body instanceof FormUrlEncodedHttpRequestBody) {
-                ((FormUrlEncodedHttpRequestBody) body).getData().forEach((name, value) -> builder.addQueryParameter(name, value));
+                ((FormUrlEncodedHttpRequestBody) body).getData().forEach((name, value) -> builder.addQueryParam(name, value));
             }
         });
         final Request build = builder.build();
@@ -84,13 +80,13 @@ final class NingHttpClientAdapterImpl implements NingHttpClientAdapter {
     }
 
     /**
-     * Creates a {@link CompletionStage} from a {@link ListenableFuture}.
+     * Creates a {@link CompletableFuture} from a {@link ListenableFuture}.
      * @param listenableFuture the future of the ning library
      * @param executor the executor to run the future in
      * @param <T> Type of the value that will be returned.
      * @return the Java 8 future implementation
      */
-    private static <T> CompletionStage<T> wrap(final ListenableFuture<T> listenableFuture, final Executor executor) {
+    private static <T> CompletableFuture<T> wrap(final ListenableFuture<T> listenableFuture, final Executor executor) {
         final CompletableFuture<T> result = new CompletableFuture<>();
         final Runnable listener = () -> {
             try {
@@ -104,7 +100,7 @@ final class NingHttpClientAdapterImpl implements NingHttpClientAdapter {
         return result;
     }
 
-    private CompletionStage<Response> wrap(final ListenableFuture<Response> listenableFuture) {
+    private CompletableFuture<Response> wrap(final ListenableFuture<Response> listenableFuture) {
         return wrap(listenableFuture, threadPool);
     }
 }
