+++ sphere-java-client-apache-async/src/it/java/io/sphere/sdk/client/ApacheHttpClientAdapterTest.java
@@ -1,5 +1,6 @@
 package io.sphere.sdk.client;
 
+import io.sphere.sdk.http.ApacheHttpClientAdapter;
 import io.sphere.sdk.http.HttpClient;
 
 public class ApacheHttpClientAdapterTest extends HttpClientAdapterTest {
