+++ sphere-java-client-ning-1_8/src/it/java/io/sphere/sdk/client/NingHttpClientAdapterTest.java
@@ -1,6 +1,7 @@
 package io.sphere.sdk.client;
 
 import io.sphere.sdk.http.HttpClient;
+import io.sphere.sdk.http.NingHttpClientAdapter;
 
 public class NingHttpClientAdapterTest extends HttpClientAdapterTest {
     @Override
