+++ sphere-java-client-ning-1_8/src/main/java/io/sphere/sdk/client/SphereAccessTokenSupplierFactory.java
@@ -1,6 +1,7 @@
 package io.sphere.sdk.client;
 
 import io.sphere.sdk.http.HttpClient;
+import io.sphere.sdk.http.NingHttpClientAdapter;
 
 public final class SphereAccessTokenSupplierFactory {
     private SphereAccessTokenSupplierFactory() {
