+++ sphere-models/src/it/java/io/sphere/sdk/client/SphereConstantAccessTokenSupplierImplTest.java
@@ -1,6 +1,7 @@
 package io.sphere.sdk.client;
 
 import io.sphere.sdk.categories.queries.CategoryQuery;
+import io.sphere.sdk.http.NingHttpClientAdapter;
 import io.sphere.sdk.test.IntegrationTest;
 import org.junit.Test;
